xe: jit: introduce hw_t to the stable IR interface
