+++ src/gpu/intel/jit/ir/hw.hpp
@@ -20,6 +20,7 @@
 #include "gpu/intel/compute/device_info.hpp"
 #include "gpu/intel/engine.hpp"
 #include "gpu/intel/jit/ir/core.hpp"
+#include "gpu/intel/jit/ir/include/hw.hpp"
 #include "gpu/intel/jit/utils/ngen_type_bridge.hpp"
 #include "gpu/intel/jit/utils/utils.hpp"
 #include "gpu/intel/utils.hpp"
@@ -29,92 +30,6 @@ namespace impl {
 namespace gpu {
 namespace intel {
 namespace jit {
-
-// Provides access to HW configuration which includes non-configurable
-// properties.
-
-namespace hw {
-enum class attr_t { none = 0, large_grf = 1, systolic = 2, atomic_fp64 = 4 };
-constexpr attr_t operator&(attr_t a, attr_t b) {
-    return static_cast<attr_t>(
-            static_cast<uint32_t>(a) & static_cast<uint32_t>(b));
-}
-constexpr attr_t operator|(attr_t a, attr_t b) {
-    return static_cast<attr_t>(
-            static_cast<uint32_t>(a) | static_cast<uint32_t>(b));
-}
-constexpr attr_t operator~(attr_t a) {
-    return static_cast<attr_t>(~static_cast<uint32_t>(a));
-}
-constexpr bool any(attr_t a) {
-    return a != static_cast<attr_t>(0);
-}
-
-inline attr_t &operator|=(attr_t &a, attr_t b) {
-    return a = a | b;
-}
-inline attr_t &operator&=(attr_t &a, attr_t b) {
-    return a = a & b;
-}
-
-} // namespace hw
-
-class hw_t {
-public:
-    using attr_t = hw::attr_t;
-    hw_t() = default;
-    explicit hw_t(const ngen::Product &product, int eu_count, int max_wg_size,
-            size_t l3_cache_size, attr_t attr);
-    ngen::HW ngen_hw() const { return hw_; }
-    operator ngen::HW() const { return hw_; }
-    const ngen::Product &product() const;
-
-    bool is_undef() const { return hw_ == ngen::HW::Unknown; }
-    bool has_fp64_atomic_support() const {
-        return any(attr_ & attr_t::atomic_fp64);
-    }
-    ngen::ProductFamily family() const;
-    int stepping() const;
-    int eu_count() const { return eu_count_; }
-    int large_grf_support() const { return any(attr_ & attr_t::large_grf); }
-    int grf_size() const;
-    int systolic_support() const { return any(attr_ & attr_t::systolic); }
-    size_t l3_cache_size() const { return l3_cache_size_; }
-
-    int max_tg_size(int regs, int simd) const;
-    // Number of EUs per Xe core (maps to dual subslice on XeHPG).
-    int eus_per_core() const;
-    int threads_per_eu(int regs = 128) const;
-    int cache_line_size() const;
-    std::string str() const;
-
-    std::string brief_str() const;
-
-    IR_DEFINE_DUMP()
-
-    bool operator<(ngen::HW rhs) const { return hw_ < rhs; }
-    bool operator>(ngen::HW rhs) const { return hw_ > rhs; }
-    bool operator<=(ngen::HW rhs) const { return hw_ <= rhs; }
-    bool operator>=(ngen::HW rhs) const { return hw_ >= rhs; }
-    bool operator==(ngen::HW rhs) const { return hw_ == rhs; }
-    bool operator!=(ngen::HW rhs) const { return hw_ != rhs; }
-#if __cplusplus >= 202002L
-    bool operator==(const hw_t &other) const = default;
-#endif
-
-private:
-    int max_wg_size(int regs = 128) const {
-        bool is_large_grf = (regs > 128);
-        return is_large_grf ? max_wg_size_ / 2 : max_wg_size_;
-    }
-
-    ngen::HW hw_ = ngen::HW::Unknown;
-    ngen::Product product_ = {};
-    int eu_count_ = 0;
-    int max_wg_size_ = 0;
-    size_t l3_cache_size_ = 0;
-    attr_t attr_ = attr_t::none;
-};
 
 inline hw_t make_ir_hw(const impl::engine_t *engine) {
     using namespace compute;
